--- demos/stitching/README.md
@@ -1,8 +1,0 @@
-### Quick and dirty demo of how to do photo stitching in command line using Python's `opencv-python
-
-not gonna try and configure this to work with our frontend and backend yet, since idk what majd is doing
-
-basically when you run stitch.py you get prompted to input the path to a directory (containing the images to be stitched)
-and then the output gets saved as a separate image file.
-
-happens pretty quick: took abt 10 seconds to stitch 16 12MP images into a 34MP panorama on my macbook
